[ticket/11] Improve scrutinizer code quality by introducing helper
